+++ src/app/(portal)/layout.tsx
@@ -22,7 +22,7 @@ export default async function PortalLayout({
         <AssumeUserBanner />
         <div className="min-h-screen bg-gray-50 flex font-sans text-gray-900">
           <Sidebar />
-          <main className="flex-1 ml-0 lg:ml-64 p-4 pt-16 lg:pt-8 lg:pr-8 lg:pb-8 lg:pl-12">
+          <main className="flex-1 min-w-0 ml-0 lg:ml-64 p-4 pt-16 lg:pt-8 lg:pr-8 lg:pb-8 lg:pl-12 overflow-x-hidden">
             {children}
           </main>
         </div>